refactor: rename ic to ir
