+++ src/lib.rs
@@ -8,7 +8,7 @@ pub mod codegen;
 pub mod column;
 pub mod database;
 pub mod expr;
-pub mod ic;
+pub mod ir;
 pub mod identifier;
 pub mod parser;
 pub mod schema;
@@ -18,7 +18,7 @@ pub mod vm;
 
 pub use column::Column;
 pub use database::Database;
-pub use ic::{Instruction, IntermediateCode};
+pub use ir::{Instruction, IntermediateCode};
 pub use identifier::BoundedString;
 pub use table::Table;
 pub use value::Value;
